+++ packages/drivers/odsp-socket-storage/src/index.ts
@@ -19,3 +19,4 @@ export * from "./odspDocumentDeltaConnection";
 export * from "./odspCache";
 export * from "./createFile";
 export * from "./odspUrlHelper";
+export * from "./decodeOdspUrl";
