+++ packages/drivers/odsp-socket-storage/src/odspDriverUrlResolver.ts
@@ -13,6 +13,7 @@ import {
 import { IOdspResolvedUrl } from "./contracts";
 import { getHashedDocumentId, INewFileInfoHeader } from "./odspUtils";
 import { getApiRoot } from "./odspUrlHelper";
+import { decodeOdspUrl } from "./decodeOdspUrl";
 
 function getSnapshotUrl(siteUrl: string, driveId: string, itemId: string) {
     const siteOrigin = new URL(siteUrl).origin;
@@ -29,35 +30,6 @@ function removeBeginningSlash(str: string): string {
     }
 
     return str;
-}
-
-export function decodeOdspUrl(url: string): { siteUrl: string; driveId: string; itemId: string; path: string } {
-    const [siteUrl, queryString] = url.split("?");
-
-    const searchParams = new URLSearchParams(queryString);
-
-    const driveId = searchParams.get("driveId");
-    const itemId = searchParams.get("itemId");
-    const path = searchParams.get("path");
-
-    if (driveId === null) {
-        throw new Error("ODSP URL did not contain a drive id");
-    }
-
-    if (itemId === null) {
-        throw new Error("ODSP Url did not contain an item id");
-    }
-
-    if (path === null) {
-        throw new Error("ODSP Url did not contain a path");
-    }
-
-    return {
-        siteUrl,
-        driveId: decodeURIComponent(driveId),
-        itemId: decodeURIComponent(itemId),
-        path: decodeURIComponent(path),
-    };
 }
 
 export class OdspDriverUrlResolver implements IUrlResolver {
